+++ src/ScottPlot5/ScottPlot5 Demos/ScottPlot5 WinForms Demo/Program.cs
@@ -11,7 +11,7 @@ static class Program
         Application.EnableVisualStyles();
 
         // CTRL+D opens this window (useful for testing in development)
-        Type testingFormType = typeof(Demos.CustomFonts);
+        Type testingFormType = typeof(Demos.CustomAxis);
 
         Application.Run(new MainMenuForm(testingFormType));
     }